other(test): make the relocated optimum sampler tests importable
The sampler tests now live in tests/optimum/v1/sample/, so add the
__init__.py that makes it a package -- without it pytest imports
test_sampler.py as a top-level module and its `from .utils import ...`
fails outright -- and split the helpers and fixtures they need out of
tests/optimum/v1/worker/.

sample/utils.py takes the sampler-only helpers. fake_load_model and the
config constants stay in worker/utils.py and are imported from there:
NUM_BLOCKS is derived from those constants and MockModelWrapper needs it,
so moving them to sample/ would invert the dependency. make_kv_cache_config,
finish_request and create_grammar_bitmask had no callers left and are gone,
which also drops the xgrammar LazyLoader.

worker/conftest.py keeps only the two fixtures its own tests need. Its
tests never reach execute_model or sample_tokens, and prepare_rbln_sampler
allocates CPU tensors without compiling, so the inductor-cache reset and
the per-xdist-worker device pinning belong with the sampler tests, which
are the ones that run -n 8 across eight NPUs.

That leaves skip_sync_vllm_and_optimum and set_npu_env_var duplicated
across core/, worker/ and sample/. Hoisting them to a v1/ conftest is not
an option: it would also catch v1/models/, which runs real models on a
real NPU that may be a CA22, and RBLN_FORCE_NPU_NAME=RBLN-CA25 plus a
_resolve_rbln_config patched to None would both lie to it.

Optimum path only.
