+++ tests/optimum/v1/sample/__init__.py
@@ -11,23 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from unittest.mock import patch
-
-import pytest
-
-
-@pytest.fixture(autouse=True)
-def skip_sync_vllm_and_optimum():
-    # Force `compiled_rbln_config = None` so sync_vllm_and_optimum() takes the
-    # sync_from_vllm() path instead of resolving a real compiled config.
-    with patch(
-        "vllm_rbln.utils.optimum.converter.dispatch._resolve_rbln_config",
-        return_value=None,
-    ):
-        yield
-
-
-@pytest.fixture(autouse=True)
-def set_npu_env_var(monkeypatch):
-    monkeypatch.setenv("RBLN_FORCE_NPU_NAME", "RBLN-CA25")
